Detect Firefox profiles already running

+++ .bashrc.d/37-browser.sh
@@ -6,8 +6,12 @@ if [[ "$OSTYPE" == 'darwin'* ]]; then
         # Prevent error message, "A copy of Firefox is already open. Only one
         # copy of Firefox can be open at a time." More info:
         # - https://apple.stackexchange.com/a/238169
-        rm "$HOME/Library/Application Support/Firefox/Profiles/"*"/.parentlock" 2>/dev/null
-        /Applications/Firefox.app/Contents/MacOS/firefox --no-remote -P "$@" &
+        if ! pgrep -f "firefox.*-P $1" >/dev/null; then
+            rm "$HOME/Library/Application Support/Firefox/Profiles/"*"/.parentlock" 2>/dev/null
+            /Applications/Firefox.app/Contents/MacOS/firefox --no-remote -P "$1" &
+        else
+            echo "Firefox already running with profile $1"
+        fi
     }
     fpa()  # Firefox profile _all_.
     {
